+++ apps/server/__tests__/service/env.test.ts
@@ -1,4 +1,4 @@
-import { testEnv } from './utils';
+import { testEnv } from '$/__tests__/utils';
 
 jest.mock('dotenv');
 
@@ -20,7 +20,7 @@ describe('env', () => {
     process.env.SMTP_FROM = 'from';
 
     // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const env = require('../service/env');
+    const env = require('../../service/env');
     expect(env.JWT_SECRET).toBe('test-secret');
     expect(env.SERVER_PORT).toBe(123);
     expect(env.URL_PORT).toBe(':123');
@@ -52,7 +52,7 @@ describe('env', () => {
     delete process.env.SMTP_FROM;
 
     // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const env = require('../service/env');
+    const env = require('../../service/env');
     expect(env.JWT_SECRET).toBe('');
     expect(env.SERVER_PORT).toBe(8080);
     expect(env.URL_PORT).toBe(':8080');
@@ -73,7 +73,7 @@ describe('env', () => {
     process.env.SERVER_PORT = '80';
 
     // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const env = require('../service/env');
+    const env = require('../../service/env');
     expect(env.URL_PORT).toBe('');
   });
 
@@ -82,7 +82,7 @@ describe('env', () => {
     process.env.SMTP_USER = 'user';
 
     // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const env = require('../service/env');
+    const env = require('../../service/env');
     expect(env.SMTP_FROM).toBe('user');
   });
 });
